+++ common/src/main/java/fr/tathan/sky_aesthetics/client/skies/settings/CloudSettings.java
@@ -7,7 +7,6 @@ import java.util.Optional;
 
 public record CloudSettings(boolean showCloud, Integer cloudHeight) {
 
-
     public static CloudSettings createDefaultSettings() {
         return new CloudSettings(true, 192);
     }