+++ changelog.md
@@ -1,4 +1,8 @@
-# Sky Aesthetics 2.0.8-BETA Changelog
+# Sky Aesthetics 2.0.9-BETA Changelog
 
-## Changes
-- Update exo config to fix a crash when joining a world.
+## Features
+- Added new render condition : player height
+- The Skybox can now be rotated (see the wiki for more information)
+
+## Change
+- Change default skybox from saturn to milky way (- 7Mb of size 💀)